options/ansi: Implement mbstowcs()

+++ options/ansi/include/stdlib.h
@@ -104,8 +104,8 @@ int wctomb(char *mb_chr, wchar_t wc);
 
 // [7.22.8] Multibyte string conversion functions
 
-int mbstowcs(wchar_t *__restrict wc_string, const char *__restrict mb_string, size_t max_size);
-int wcstombs(char *mb_string, const wchar_t *__restrict wc_string, size_t max_size);
+size_t mbstowcs(wchar_t *__restrict wc_string, const char *__restrict mb_string, size_t max_size);
+size_t wcstombs(char *mb_string, const wchar_t *__restrict wc_string, size_t max_size);
 
 // posix extension
